Rework SeriesAdapter to extend native RecyclerView.Adapter
Reworked SeriesAdapter to extend the native RecyclerView.Adapter
instead of RealmRecyclerView.Adapter. This is done in order to allow
removing Series from the Adapter's data set, which cannot be done with
the Realm version. Because of this change, methods were added to
manually update the data set when data changes are detected (from
syncing). The Spinner in the Toolbar of SeasonsFragment is now
correctly restored on configuration changes.